+++ src/ft_nbr/ft_itoa_base.c
@@ -6,7 +6,7 @@
 /*   By: dpoveda- <me@izenynn.com>                  +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2021/10/21 01:53:11 by dpoveda-          #+#    #+#             */
-/*   Updated: 2023/03/13 17:25:54 by dpoveda-         ###   ########.fr       */
+/*   Updated: 2023/03/13 19:33:56 by dpoveda-         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -22,14 +22,14 @@ char	*ft_itoa_base(int n, const char *base)
 	char	*out;
 	short	sign;
 	short	is_neg;
-	int		base_l;
+	size_t	base_l;
 
 	is_neg = 0;
 	sign = 1;
 	if (n < 0 && ++is_neg)
 		sign = -1;
 	base_l = ft_strlen(base);
-	out = (char *)ft_calloc(1, ft_intlen_base(n, base_l) + is_neg + 1);
+	out = (char *)ft_calloc(1, (size_t)ft_intlen_base(n, (int)base_l) + (size_t)is_neg + 1);
 	if (!out)
 		return (NULL);
 	i = 0;
@@ -37,8 +37,8 @@ char	*ft_itoa_base(int n, const char *base)
 		out[i++] = '0';
 	while (n)
 	{
-		out[i++] = base[n % base_l * sign];
-		n /= base_l;
+		out[i++] = base[n % (int)base_l * sign];
+		n /= (int)base_l;
 	}
 	if (sign == -1)
 		out[i++] = '-';
